# MNT Speedup example plot_select_from_model_diabetes.py