--- a/examples/feature_selection/plot_select_from_model_diabetes.py
+++ b/examples/feature_selection/plot_select_from_model_diabetes.py
@@ -29,10 +29,6 @@ from sklearn.datasets import load_diabetes
 
 diabetes = load_diabetes()
 X, y = diabetes.data, diabetes.target
-# Using a subset of columns to demo
-columns_interest = ["sex", "bmi", "s1", "s5", "s6"]
-feat_idx = dict(list(zip(diabetes.feature_names, range(len(diabetes.feature_names)))))
-X = X[:, [feat_idx[feature] for feature in columns_interest]]
 print(diabetes.DESCR)
 
 # %%
@@ -50,11 +46,11 @@ print(diabetes.DESCR)
 # :ref:`sphx_glr_auto_examples_inspection_plot_linear_model_coefficient_interpretation.py`.
 import matplotlib.pyplot as plt
 import numpy as np
-from sklearn.linear_model import LassoCV
+from sklearn.linear_model import RidgeCV
 
-lasso = LassoCV().fit(X, y)
-importance = np.abs(lasso.coef_)
-feature_names = np.array(columns_interest)
+ridge = RidgeCV(alphas=np.logspace(-6, 6, num=5)).fit(X, y)
+importance = np.abs(ridge.coef_)
+feature_names = np.array(diabetes.feature_names)
 plt.bar(height=importance, x=feature_names)
 plt.title("Feature importances via coefficients")
 plt.show()
@@ -77,7 +73,7 @@ from time import time
 threshold = np.sort(importance)[-3] + 0.01
 
 tic = time()
-sfm = SelectFromModel(lasso, threshold=threshold).fit(X, y)
+sfm = SelectFromModel(ridge, threshold=threshold).fit(X, y)
 toc = time()
 print(f"Features selected by SelectFromModel: {feature_names[sfm.get_support()]}")
 print(f"Done in {toc - tic:.3f}s")
@@ -102,13 +98,13 @@ from sklearn.feature_selection import SequentialFeatureSelector
 
 tic_fwd = time()
 sfs_forward = SequentialFeatureSelector(
-    lasso, n_features_to_select=2, direction="forward"
+    ridge, n_features_to_select=2, direction="forward"
 ).fit(X, y)
 toc_fwd = time()
 
 tic_bwd = time()
 sfs_backward = SequentialFeatureSelector(
-    lasso, n_features_to_select=2, direction="backward"
+    ridge, n_features_to_select=2, direction="backward"
 ).fit(X, y)
 toc_bwd = time()
 
@@ -131,8 +127,10 @@ print(f"Done in {toc_bwd - tic_bwd:.3f}s")
 # features. In general, this isn't the case and the two methods would lead to
 # different results.
 #
-# We also note that the features selected by SFS are the same as the features
-# selected by feature importance. It is quite remarkable considering
+# We also note that the features selected by SFS differ from those selected by
+# feature importance: SFS selects `bmi` instead of `s1`. This does sounds
+# reasonable though, since `bmi` corresponds to the third most important
+# feature according to the coefficients. It is quite remarkable considering
 # that SFS makes no use of the coefficients at all.
 #
 # To finish with, we should note that