<!--
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/scikit-learn/scikit-learn/blob/main/CONTRIBUTING.md
-->

#### Reference Issues/PRs
<!--
Example: Fixes #1234. See also #3456.
Please use keywords (e.g., Fixes) to create link to the issues or pull requests
you resolved, so that they will automatically be closed when your pull request
is merged. See https://github.com/blog/1506-closing-issues-via-pull-requests
-->
#21598

#### What does this implement/fix? Explain your changes.
Speeds up plot_select_model_diabetes.py by using a subset of features while preserving the original message of the example. The major bottleneck, in this case, was the backward SFS which could be sped up using fewer features. 

```
Features selected by SelectFromModel: ['s1' 's5']
Done in 0.046s
Features selected by forward sequential selection: ['bmi' 's5']
Done in 2.907s
Features selected by backward sequential selection: ['bmi' 's5']
Done in 8.544s
real 15.43
user 13.27
sys 0.96
```
![original_diabetes](https://user-images.githubusercontent.com/28287747/142781904-2683b67a-4573-498c-a79b-372385945e1c.png)
```

Features selected by SelectFromModel: ['bmi' 's5']
Done in 0.033s
Features selected by forward sequential selection: ['bmi' 's5']
Done in 1.407s
Features selected by backward sequential selection: ['bmi' 's5']
Done in 1.848s
real 6.87
user 5.11
sys 0.87
```
![augmented_diabetes](https://user-images.githubusercontent.com/28287747/142781888-f8d0326f-320a-4108-916f-44296b78a017.png)

#### Any other comments?


<!--
Please be aware that we are a loose team of volunteers so patience is
necessary; assistance handling other issues is very welcome. We value
all user contributions, no matter how minor they are. If we are slow to
review, either the pull request needs some benchmarking, tinkering,
convincing, etc. or more likely the reviewers are simply busy. In either
case, we ask for your understanding during the review process.
For more information, see our FAQ on this topic:
http://scikit-learn.org/dev/faq.html#why-is-my-pull-request-not-getting-any-attention.

Thanks for contributing!
-->