@ogrisel 2021-11-22
I find it weird to do an example on automated feature selection on a dataset which was prepared with manual feature selection.

However, since the example does not really rely on the choice of the Lasso penalty, maybe we could use just a faster base linear model, for instance `RidgeCV(alphas=np.logspace(-6, 6, num=5))` on the original feature set.