@glemaitre 2021-11-22
> However, since the example does not really rely on the choice of the Lasso penalty, maybe we could use just a faster base linear model, for instance RidgeCV(alphas=np.logspace(-6, 6, num=5)) on the original feature set.

I agree with @ogrisel. In the original example, the penalty chosen will let out only a single feature. So using `RidgeCV` will be much cheaper and keep the description more or less the same. I think that what is really expensive here is indeed the `SequentialFeatureSelector`. We could reduce to `cv=3` and check if this is enough (maybe with an additional note to mention that in practice one should consider to increase the number of folds)