@ArthDh 2021-11-22
Using RidgeCV the results are much faster on the original feature set:
```
Features selected by SelectFromModel: ['s1' 's5']
Done in 0.001s
Features selected by forward sequential selection: ['bmi' 's5']
Done in 0.116s
Features selected by backward sequential selection: ['bmi' 's5']
Done in 0.305s
real 6.60
user 2.20
sys 0.88
```
Graph generated:
![ridge](https://user-images.githubusercontent.com/28287747/142934406-973dab46-1928-46b5-bf04-8b68ce9c8f0f.png)

Should I move forward and update the file?