@glemaitre 2021-11-23
> Using RidgeCV the results are much faster on the original feature set:

Yes this is expected. Can you make the change?